Migrate matrices to Eigen, sizes to size_t, add checked narrowing
mat_type<T> now derives from Eigen::Matrix<T,Dynamic,Dynamic> (Eigen's
sanctioned extension pattern), so stored matrices participate directly
in Eigen expressions while keeping the historical semantics the tests
pin: zero-initializing sized ctor/resize, unsigned->size_t num_rows/
num_cols, storage-order (column-major) linear operator[] and iterators,
bounds-checked operator()(i,j), get_row/get_column copy-outs. Eigen
3.4.0 is fetched headers-only and linked PUBLIC; the wire format is
unchanged (Eigen's column-major default matches the old layout, golden
bytes still pass).

All public size types (get_num_elements, num_rows/cols, operator[],
operator(), get_*_at, set_vector/set_vec_*/set_mat_*) move from
unsigned to std::size_t; interface layers and the toml/nlohmann/lua
counters are adjusted.

Numeric narrowing is now range-checked everywhere: integer->integer via
std::in_range, real/complex->integer via GC_details::real_fits_integral
(exact power-of-two bounds, strict upper comparison; NaN/Inf rejected).
This closes the silent LONG->int truncation in get_value<int_type> and
the unchecked isInteger()-only casts (1e300 was UB). GC_DO_ERROR now
funnels through a free gc_do_error() so mat_type (defined before
GenericContainer) can use it; GenericContainer::exception stays.

Conversion-check regressions un-skipped: 1672 assertions green,
164/164 CTest, ASan+UBSan clean.

+++ src/GenericContainerSerialize.cc
@@ -550,7 +550,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vec_pointer( static_cast<unsigned>( i32 ) );
+        allocate_vec_pointer( static_cast<std::size_t>( i32 ) );
         for ( auto & p : _v_p() )
         {
           int64_t ptr_value{ 0 };
@@ -581,7 +581,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vec_int( static_cast<unsigned>( i32 ) );
+        allocate_vec_int( static_cast<std::size_t>( i32 ) );
         for ( auto & i : _v_i() )
         {
           sz = buffer_to_int32( buffer, buffer_dim - nbyte, &i );
@@ -594,7 +594,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vec_long( static_cast<unsigned>( i32 ) );
+        allocate_vec_long( static_cast<std::size_t>( i32 ) );
         for ( auto & i : _v_l() )
         {
           sz = buffer_to_int64( buffer, buffer_dim - nbyte, &i );
@@ -607,7 +607,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vec_real( static_cast<unsigned>( i32 ) );
+        allocate_vec_real( static_cast<std::size_t>( i32 ) );
         for ( auto & r : _v_r() )
         {
           sz = buffer_to_double( buffer, buffer_dim - nbyte, &r );
@@ -620,7 +620,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vec_complex( static_cast<unsigned>( i32 ) );
+        allocate_vec_complex( static_cast<std::size_t>( i32 ) );
         for ( auto & c : _v_c() )
         {
           sz = buffer_to_double( buffer, buffer_dim - nbyte, &bf );
@@ -641,7 +641,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( nr >= 0 && nc >= 0, "GenericContainer::de_serialize, invalid matrix dimensions" );
-        allocate_mat_int( static_cast<unsigned>( nr ), static_cast<unsigned>( nc ) );
+        allocate_mat_int( static_cast<std::size_t>( nr ), static_cast<std::size_t>( nc ) );
         for ( auto & i : _m_i() )
         {
           sz = buffer_to_int32( buffer, buffer_dim - nbyte, &i );
@@ -657,7 +657,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( nr >= 0 && nc >= 0, "GenericContainer::de_serialize, invalid matrix dimensions" );
-        allocate_mat_long( static_cast<unsigned>( nr ), static_cast<unsigned>( nc ) );
+        allocate_mat_long( static_cast<std::size_t>( nr ), static_cast<std::size_t>( nc ) );
         for ( auto & i : _m_l() )
         {
           sz = buffer_to_int64( buffer, buffer_dim - nbyte, &i );
@@ -673,7 +673,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( nr >= 0 && nc >= 0, "GenericContainer::de_serialize, invalid matrix dimensions" );
-        allocate_mat_real( static_cast<unsigned>( nr ), static_cast<unsigned>( nc ) );
+        allocate_mat_real( static_cast<std::size_t>( nr ), static_cast<std::size_t>( nc ) );
         for ( auto & r : _m_r() )
         {
           sz = buffer_to_double( buffer, buffer_dim - nbyte, &r );
@@ -689,7 +689,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( nr >= 0 && nc >= 0, "GenericContainer::de_serialize, invalid matrix dimensions" );
-        allocate_mat_complex( static_cast<unsigned>( nr ), static_cast<unsigned>( nc ) );
+        allocate_mat_complex( static_cast<std::size_t>( nr ), static_cast<std::size_t>( nc ) );
         for ( auto & c : _m_c() )
         {
           sz = buffer_to_double( buffer, buffer_dim - nbyte, &bf );
@@ -707,7 +707,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vec_string( static_cast<unsigned>( i32 ) );
+        allocate_vec_string( static_cast<std::size_t>( i32 ) );
         for ( auto & s : _v_s() )
         {
           nb = buffer_to_int32( buffer, buffer_dim - nbyte, &i32 );
@@ -725,7 +725,7 @@ namespace GC_namespace
         buffer += nb;
         nbyte += nb;
         GC_ASSERT( i32 >= 0, "GenericContainer::de_serialize, invalid vector size" );
-        allocate_vector( static_cast<unsigned>( i32 ) );
+        allocate_vector( static_cast<std::size_t>( i32 ) );
         for ( auto & S : _v() )
         {
           sz = S.de_serialize( buffer_dim - nbyte, buffer );
